rename zos_traits tp api
